--- server/src/main/java/com/flipkart/security/grayskull/configuration/properties/ReadOnlyAppProperties.java
@@ -1,14 +1,0 @@
-package com.flipkart.security.grayskull.configuration.properties;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Setter
-@ConfigurationProperties(prefix = "server.read-only")
-@Component
-public class ReadOnlyAppProperties {
-    private boolean enabled;
-}
