--- simple-app/src/main/java/com/flipkart/security/grayskull/GrayskullApplication.java
@@ -1,13 +1,0 @@
-package com.flipkart.security.grayskull;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GrayskullApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GrayskullApplication.class, args);
-    }
-
-}
