+++ simple-app/src/test/java/com/flipkart/grayskull/app/TestGrayskullApplication.java
@@ -1,0 +1,11 @@
+package com.flipkart.grayskull.app;
+
+import org.springframework.boot.SpringApplication;
+
+public class TestGrayskullApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.from(GrayskullApplication::main).with(TestcontainersConfiguration.class).run(args);
+    }
+
+}
