+++ simple-app/src/test/java/com/flipkart/grayskull/app/TestcontainersConfiguration.java
@@ -1,0 +1,8 @@
+package com.flipkart.grayskull.app;
+
+import org.springframework.boot.test.context.TestConfiguration;
+
+@TestConfiguration(proxyBeanMethods = false)
+class TestcontainersConfiguration {
+
+}
