--- simple-app/src/test/java/com/flipkart/security/grayskull/GrayskullApplicationTests.java
@@ -1,15 +1,0 @@
-package com.flipkart.security.grayskull;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-
-@Import(TestcontainersConfiguration.class)
-@SpringBootTest
-class GrayskullApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
