--- simple-app/src/test/java/com/flipkart/security/grayskull/TestcontainersConfiguration.java
@@ -1,8 +1,0 @@
-package com.flipkart.security.grayskull;
-
-import org.springframework.boot.test.context.TestConfiguration;
-
-@TestConfiguration(proxyBeanMethods = false)
-class TestcontainersConfiguration {
-
-}
