--- simple-impl/src/main/java/com/flipkart/security/grayskull/crypto/EncryptionDto.java
@@ -1,17 +1,0 @@
-package com.flipkart.security.grayskull.crypto;
-
-import jakarta.validation.constraints.NotEmpty;
-
-public class EncryptionDto {
-
-    @NotEmpty
-    private String encryptionKey;
-
-    public String getEncryptionKey() {
-        return encryptionKey;
-    }
-
-    public void setEncryptionKey(String encryptionKey) {
-        this.encryptionKey = encryptionKey;
-    }
-}
