--- simple-impl/src/main/java/com/flipkart/security/grayskull/crypto/KeyProperties.java
@@ -1,24 +1,0 @@
-package com.flipkart.security.grayskull.crypto;
-
-import jakarta.validation.constraints.NotEmpty;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.validation.annotation.Validated;
-
-import java.util.Map;
-
-@Configuration
-@Validated
-@ConfigurationProperties(prefix = "grayskull.crypto")
-public class KeyProperties {
-    @NotEmpty
-    private Map<String, String> keys;
-
-    public Map<String, String> getKeys() {
-        return keys;
-    }
-
-    public void setKeys(Map<String, String> keys) {
-        this.keys = keys;
-    }
-}
