--- spi/src/main/java/com/flipkart/security/grayskull/spi/EncryptionService.java
@@ -1,6 +1,0 @@
-package com.flipkart.security.grayskull.spi;
-
-public interface EncryptionService {
-    byte[] encrypt(byte[] data, String keyId);
-    byte[] decrypt(byte[] data, String keyId);
-}
